+++ test/libsolidity/semanticTests/array/byte_array_storage_layout.sol
@@ -2,38 +2,38 @@ contract c {
     bytes data;
     function test_short() public returns (uint256 r) {
         assembly {
-            sstore(data_slot, 0)
+            sstore(data.slot, 0)
         }
         for (uint8 i = 0; i < 15; i++) {
             data.push(bytes1(i));
         }
         assembly {
-            r := sload(data_slot)
+            r := sload(data.slot)
         }
     }
 
     function test_long() public returns (uint256 r) {
         assembly {
-            sstore(data_slot, 0)
+            sstore(data.slot, 0)
         }
         for (uint8 i = 0; i < 33; i++) {
             data.push(bytes1(i));
         }
         assembly {
-            r := sload(data_slot)
+            r := sload(data.slot)
         }
     }
 
     function test_pop() public returns (uint256 r) {
         assembly {
-            sstore(data_slot, 0)
+            sstore(data.slot, 0)
         }
         for (uint8 i = 0; i < 32; i++) {
             data.push(bytes1(i));
         }
         data.pop();
         assembly {
-            r := sload(data_slot)
+            r := sload(data.slot)
         }
     }
 }