+++ test/libsolidity/semanticTests/libraries/library_delegatecall_guard_pure.sol
@@ -12,13 +12,13 @@ contract C {
     }
     function g() public returns (bool, uint256) {
         uint256 ys;
-        assembly { ys := y_slot }
+        assembly { ys := y.slot }
         (bool success, bytes memory data) = address(L).delegatecall(abi.encodeWithSelector(L.f.selector, ys));
         return (success, success ? abi.decode(data,(uint256)) : 0);
     }
     function h() public returns (bool, uint256) {
         uint256 ys;
-        assembly { ys := y_slot }
+        assembly { ys := y.slot }
         (bool success, bytes memory data) = address(L).call(abi.encodeWithSelector(L.f.selector, ys));
         return (success, success ? abi.decode(data,(uint256)) : 0);
     }