+++ test/libsolidity/semanticTests/viaYul/array_storage_index_zeroed_test.sol
@@ -20,7 +20,7 @@ contract C {
             for (uint i = 3; i < len; i++)
             {
                 assembly {
-                    mstore(0, storageArray_slot)
+                    mstore(0, storageArray.slot)
                     let pos := add(keccak256(0, 0x20), i)
 
                     if iszero(eq(sload(pos), 0)) {