+++ test/libsolidity/semanticTests/viaYul/storage/dirty_storage_static_array.sol
@@ -2,7 +2,7 @@ contract C {
     uint8[1] s;
     function f() public returns (bool correct) {
         assembly {
-            sstore(s_slot, 257)
+            sstore(s.slot, 257)
         }
         uint8 x = s[0];
         uint r;