+++ test/libsolidity/semanticTests/viaYul/storage/dirty_storage_struct.sol
@@ -6,7 +6,7 @@ contract C {
     function f() public returns (bool correct) {
         s.m.push();
         assembly {
-            mstore(0, s_slot)
+            mstore(0, s.slot)
             sstore(keccak256(0, 0x20), 257)
         }
         uint8 x = s.m[0];