+++ test/libsolidity/syntaxTests/controlFlow/storageReturn/assembly/switch_fine.sol
@@ -4,15 +4,15 @@ contract C {
     function f(bool flag) internal pure returns (S storage c) {
         assembly {
             switch flag
-            case 0 { c_slot := s_slot }
-            default { c_slot := s_slot }
+            case 0 { c.slot := s.slot }
+            default { c.slot := s.slot }
         }
     }
     function g(uint256 a) internal pure returns (S storage c) {
         assembly {
             switch a
             case 0 { revert(0, 0) }
-            default { c_slot := s_slot }
+            default { c.slot := s.slot }
         }
     }
 }